requirements.txt
pyTelegramBotAPI==4.29.1
python-dotenv==1.0.1
schedule==1.2.2
